feat: move order book container to app to enable sidebar integration
