+++ apps/trading/components/orderbook-container/index.ts
@@ -1,0 +1,1 @@
+export * from './orderbook-container';
