+++ libs/market-depth/src/lib/index.ts
@@ -1,7 +1,6 @@
 export * from './__generated__/MarketDepth';
 export * from './depth-chart';
 export * from './market-depth-provider';
-export * from './orderbook-container';
 export * from './orderbook-data';
 export * from './orderbook-manager';
 export * from './orderbook-row';
