--- libs/market-depth/src/lib/orderbook-container.tsx
@@ -1,5 +1,0 @@
-import { OrderbookManager } from './orderbook-manager';
-
-export const OrderbookContainer = ({ marketId }: { marketId: string }) => (
-  <OrderbookManager marketId={marketId} />
-);
