+++ libs/market-depth/src/lib/orderbook-manager.tsx
@@ -9,7 +9,6 @@ import type {
   MarketDepthUpdateSubscription,
   PriceLevelFieldsFragment,
 } from './__generated__/MarketDepth';
-import { useCreateOrderStore } from '@vegaprotocol/orders';
 
 export type OrderbookData = {
   asks: PriceLevelFieldsFragment[];
@@ -18,9 +17,13 @@ export type OrderbookData = {
 
 interface OrderbookManagerProps {
   marketId: string;
+  onClick?: (args: { price?: string; size?: string }) => void;
 }
 
-export const OrderbookManager = ({ marketId }: OrderbookManagerProps) => {
+export const OrderbookManager = ({
+  marketId,
+  onClick,
+}: OrderbookManagerProps) => {
   const variables = { marketId };
 
   const { data, error, loading, reload } = useDataProvider<
@@ -50,8 +53,6 @@ export const OrderbookManager = ({ marketId }: OrderbookManagerProps) => {
     dataProvider: marketDataProvider,
     variables,
   });
-  const useOrderStoreRef = useCreateOrderStore();
-  const updateOrder = useOrderStoreRef((store) => store.update);
 
   return (
     <AsyncRenderer
@@ -66,14 +67,7 @@ export const OrderbookManager = ({ marketId }: OrderbookManagerProps) => {
         decimalPlaces={market?.decimalPlaces ?? 0}
         positionDecimalPlaces={market?.positionDecimalPlaces ?? 0}
         assetSymbol={market?.tradableInstrument.instrument.product.quoteName}
-        onClick={({ price, size }) => {
-          if (price) {
-            updateOrder(marketId, { price });
-          }
-          if (size) {
-            updateOrder(marketId, { size });
-          }
-        }}
+        onClick={onClick}
         midPrice={marketData?.midPrice}
       />
     </AsyncRenderer>
